Traipse 'OpenRPG' {100203-00}
Traipse is a distribution of OpenRPG that is designed to be easy to
setup and go. Traipse also makes it easy for developers to work on code
without fear of sacrifice. 'Ornery-Orc' continues the trend of 'Grumpy'
and adds fixes to the code. 'Ornery-Orc's main goal is to offer more
advanced features and enhance the productivity of the user.

Update Summary (Stable)
New Features:
New Bookmarks Feature
New 'boot' command to remote admin
New confirmation window for sent nodes
Miniatures Layer pop up box allows users to turn off Mini labels, from
FlexiRPG
New Zoom Mouse plugin added
New Images added to Plugin UI
Switching to Element Tree
New Map efficiency, from FlexiRPG
New Status Bar to Update Manager
New TrueDebug Class in orpg_log (See documentation for usage)
New Portable Mercurial
New Tip of the Day, from Core and community
New Reference Syntax added for custom PC sheets
New Child Reference for gametree
New Parent Reference for gametree
New Gametree Recursion method, mapping, context sensitivity, and
effeciency..
New Features node with bonus nodes and Node Referencing help added
New Dieroller structure from Core
New DieRoller portability for odd Dice
New 7th Sea die roller; ie [7k3] = [7d10.takeHighest(3).open(10)]
New 'Mythos' System die roller added
New vs. die roller method for WoD; ie [3v3] = [3d10.vs(3)]. Included for
Mythos roller also
New Warhammer FRPG Die Roller (Special thanks to Puu-san for the
support)
New EZ_Tree Reference system. Push a button, Traipse the tree, get a
reference (Beta!)
New Grids act more like Spreadsheets in Use mode, with Auto Calc

Fixes:
Fix to allow for portability to an OpenSUSE linux OS
Fix to mplay_client for Fedora and OpenSUSE
Fix to Text based Server
Fix to Remote Admin Commands
Fix to Pretty Print, from Core
Fix to Splitter Nodes not being created
Fix to massive amounts of images loading, from Core
Fix to Map from gametree not showing to all clients
Fix to gametree about menus
Fix to Password Manager check on startup
Fix to PC Sheets from tool nodes. They now use the tabber_panel
Fix to Whiteboard ID to prevent random line or text deleting.
Fixes to Server, Remote Server, and Server GUI
Fix to Update Manager; cleaner clode for saved repositories
Fixes made to Settings Panel and now reactive settings when Ok is
pressed
Fixes to Alternity roller's attack roll. Uses a simple Tuple instead of
a Splice
Fix to Use panel of Forms and Tabbers. Now longer enters design mode
Fix made Image Fetching. New fetching image and new failed image
Fix to whiteboard ID's to prevent non updated clients from ruining the
fix.
default_manifest.xml renamed to default_upmana.xml

+++ orpg/chat/chat_msg.py
@@ -21,18 +21,18 @@
 # Author: Ted Berg
 # Maintainer:
 # Version:
-#   $Id: chat_msg.py,v 1.15 2006/11/04 21:24:19 digitalxero Exp $
+#   $Id: chat_msg.py,v Traipse 'Ornery-Orc' prof.ebral Exp $
 #
 # Description: Contains class definitions for manipulating <chat/> messages
 #
 #
 
-__version__ = "$Id: chat_msg.py,v 1.15 2006/11/04 21:24:19 digitalxero Exp $"
+__version__ = "$Id: chat_msg.py,v Traipse 'Ornery-Orc' prof.ebral Exp $"
 
 from orpg.orpgCore import *
 from chat_version import CHAT_VERSION
-from orpg.tools.orpg_log import logger
-from orpg.tools.decorators import debugging
+from orpg.tools.orpg_log import logger, debug
+from xml.etree.ElementTree import tostring, fromstring
 
 CHAT_MESSAGE = 1
 WHISPER_MESSAGE = 2
@@ -42,64 +42,42 @@ SYSTEM_MESSAGE = 5
 WHISPER_EMOTE_MESSAGE = 6
 
 class chat_msg:
-    @debugging
-    def __init__(self,xml_text="<chat type=\"1\" version=\""+CHAT_VERSION+"\" alias=\"\" ></chat>"):
+    
+    def __init__(self, xml_text="<chat type='1' version='"+CHAT_VERSION+"' alias='' ></chat>"):
         self.chat_dom = None
         self.takexml(xml_text)
 
-    @debugging
     def __del__(self):
-        if self.chat_dom:
-            self.chat_dom.unlink()
+        if self.chat_dom: self.chat_dom.unlink()
 
-    @debugging
     def toxml(self):
-        return component.get('xml').toxml(self.chat_dom)
+        return tostring(self.chat_dom)
 
-    @debugging
     def takexml(self,xml_text):
-        xml_dom = component.get('xml').parseXml(xml_text)
-        node_list = xml_dom.getElementsByTagName("chat")
-        if len(node_list) < 1:
-            print "Warning: no <chat/> elements found in DOM."
-        else:
-            if len(node_list) > 1:
-                print "Found more than one instance of <" + self.tagname + "/>.  Taking first one"
-            self.takedom(node_list[0])
+        xml_dom = fromstring(xml_text)
+        self.takedom(xml_dom)
 
-    @debugging
-    def takedom(self,xml_dom):
-        if self.chat_dom:
-            self.text_node = None
-            self.chat_dom.unlink()
+    def takedom(self, xml_dom):
         self.chat_dom = xml_dom
-        self.text_node = component.get('xml').safe_get_text_node(self.chat_dom)
-
-    @debugging
-    def set_text(self,text):
-        text = component.get('xml').strip_text(text)
-        self.text_node._set_nodeValue(text)
-
-    @debugging
+        self.text_node = xml_dom.text
+    
+    def set_text(self, text):
+        self.chat_dom.text = text
+    
     def set_type(self,type):
-        self.chat_dom.setAttribute("type",str(type))
-
-    @debugging
+        self.chat_dom.set("type", str(type))
+    
     def get_type(self):
-        return int(self.chat_dom.getAttribute("type"))
+        return int(self.chat_dom.get("type"))
 
-    @debugging
     def set_alias(self,alias):
-        self.chat_dom.setAttribute("alias",alias)
+        self.chat_dom.set("alias",alias)
 
-    @debugging
     def get_alias(self):
-        return self.chat_dom.getAttribute("alias")
+        return self.chat_dom.get("alias")
 
-    @debugging
     def get_text(self):
-        return self.text_node._get_nodeValue()
+        return self.text_node
 
-    @debugging
     def get_version(self):
-        return self.chat_dom.getAttribute("version")
+        return self.chat_dom.get("version")
